+++ database/migrations/2023_01_16_110229_create_asignatura_profesors_table.php
@@ -15,9 +15,9 @@ class CreateAsignaturaProfesorsTable extends Migration
     {
         Schema::create('asignatura_profesors', function (Blueprint $table) {
             $table->id();
-            $table->integer('idProfesor');
+            $table->unsignedBigInteger('idProfesor');
             $table->foreign("idProfesor")->references('id')->on('profesors')->onUpdate("cascade")->onDelete("cascade");
-            $table->integer('idAsignatura');
+            $table->unsignedBigInteger('idAsignatura');
             $table->foreign("idAsignatura")->references('id')->on('asignaturas')->onUpdate("cascade")->onDelete("cascade");
             $table->primary(["idProfesor", "idAsignatura"]);
             $table->timestamps();
